# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Currencies API",
        "description": "This service allow you getting data about currency.",
        "contact": {
            "name": "Ilya Novak",
            "email": "ilya.novak.dev@gmail.com"
        },
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/sign-in": {
            "post": {
                "tags": [
                    "auth"
                ],
                "description": "Sign in to get access token",
                "operationId": "ceb2ec20ac1dc4a1cd58334929cbe92c",
                "requestBody": {
                    "description": "Data to sign in.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "email@test.xxx"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "metadata": {
                                            "properties": {
                                                "access_token": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/sign-up": {
            "post": {
                "tags": [
                    "auth"
                ],
                "description": "Registration",
                "operationId": "31524216dd66dd05cf07ced3f7eab7f7",
                "requestBody": {
                    "description": "Data to sign in.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "test name"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "email@test.xxx"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success registration",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "metadata": {
                                            "properties": {
                                                "access_token": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/currency/{valuteId}": {
            "get": {
                "tags": [
                    "currency"
                ],
                "description": "Get info about valute by id and in range from date to date.",
                "operationId": "984a5d9c1ff55d6606969b8b0421d0e0",
                "parameters": [
                    {
                        "name": "valuteId",
                        "in": "path",
                        "description": "For example R01010, R01035."
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Date format yyyy-mm-dd"
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Date format yyyy-mm-dd"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return array for selected currency",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Currency"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Currency": {
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "valute_id": {
                        "type": "string"
                    },
                    "num_code": {
                        "type": "string"
                    },
                    "char_code": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "value": {
                        "type": "string"
                    },
                    "date": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}